+++ domain/src/main/java/com/example/ringapp/domain/usecase/RecordHabitProgressUseCase.kt
@@ -5,5 +5,5 @@ import com.example.ringapp.data.repository.HabitRepository
 import javax.inject.Inject
 
 class RecordHabitProgressUseCase @Inject constructor(private val repository: HabitRepository) {
-    suspend operator fun invoke(habit: HabitEntity, date: Long, actual: Int, note: String?) = repository.recordProgress(habit, date, actual, note)
+    suspend operator fun invoke(habit: HabitEntity, date: Long, actual: Double, note: String?) = repository.recordProgress(habit, date, actual, note)
 }
